requirements.txt
discord.py
lavalink